+++ extensions/laravel-herd/src/toggle-dumps.ts
@@ -1,13 +1,20 @@
 import { showHUD, updateCommandMetadata } from "@raycast/api";
 import { Herd } from "./utils/Herd";
+import { rescue } from "./utils/rescue";
 
 export default async function main() {
   await showHUD("Toggling Dumps");
 
-  let isIntercepting = await Herd.Dumps.isInterceptingDumps();
+  let isIntercepting = await rescue(
+    async () => Herd.Dumps.isInterceptingDumps(),
+    "Failed to check if Dumps are intercepting.",
+  );
 
-  await Herd.Dumps.toggleIntercept();
+  await rescue(async () => Herd.Dumps.toggleIntercept(), "Failed to toggle Dumps intercept.");
 
-  isIntercepting = await Herd.Dumps.isInterceptingDumps();
+  isIntercepting = await rescue(
+    async () => Herd.Dumps.isInterceptingDumps(),
+    "Failed to check if Dumps are intercepting.",
+  );
   await updateCommandMetadata({ subtitle: isIntercepting ? "Intercepting" : "Not intercepting" });
 }