+++ extensions/laravel-herd/CHANGELOG.md
@@ -1,3 +1,15 @@
 # Laravel Herd Changelog
 
-## [Initial Version] - 2025-04-08
+
+## Log Viewer Commands, Refresh & Bugfixes - {PR_MERGE_DATE}
+
+- Added command to open the log viewer for most recent site
+- Added action to open log viewer for a site in the Sites command
+- Added command to clear all Herd-related caches
+- Updated command names
+- Fixed a bug where pinning the dumps window throw an error
+- Improved error handling for Dumps commands
+- Polished README
+
+
+## [Initial Version] - April 8, 2025